+++ Sources/APIContract/Protocols/StreamingAPIContract.swift
@@ -39,6 +39,7 @@ extension StreamingAPIContract {
         var path = pathTemplate
         for (key, value) in input.pathParameters {
             path = path.replacingOccurrences(of: ":\(key)", with: value)
+            path = path.replacingOccurrences(of: "{\(key)}", with: value)
         }
         return path
     }